[#1036] Most of storages replaced with generic ones. Logic that creates ChangeInfo objects moved to static methods in `ChangeInfoFactory`.
