+++ plugins/versionpress/src/Database/wordpress-schema.yml
@@ -1,6 +1,8 @@
 post:
   table: posts
   id: ID
+  storage: 'VersionPress\Storages\PostStorage'
+  changeinfo-fn: 'VersionPress\Storages\ChangeInfoFactory::createPostChangeInfo'
   references:
     post_author: user
     post_parent: post
@@ -20,6 +22,7 @@ post:
 postmeta:
   id: meta_id
   parent-reference: post_id
+  changeinfo-fn: 'VersionPress\Storages\ChangeInfoFactory::createPostMetaChangeInfo'
   references:
     post_id: post
   value-references:
@@ -36,6 +39,8 @@ postmeta:
 comment:
   table: comments
   id: comment_ID
+  storage: 'VersionPress\Storages\CommentStorage'
+  changeinfo-fn: 'VersionPress\Storages\ChangeInfoFactory::createCommentChangeInfo'
   references:
     comment_post_ID: post
     user_id: user
@@ -46,6 +51,7 @@ comment:
 commentmeta:
   table: commentmeta
   id: meta_id
+  changeinfo-fn: 'VersionPress\Storages\ChangeInfoFactory::createCommentMetaChangeInfo'
   parent-reference: comment_id
   references:
     comment_id: comment
@@ -54,22 +60,26 @@ commentmeta:
 user:
   table: users
   id: ID
+  changeinfo-fn: 'VersionPress\Storages\ChangeInfoFactory::createUserChangeInfo'
   ignored-columns:
     - user_activation_key
 
 term:
   table: terms
   id: term_id
+  changeinfo-fn: 'VersionPress\Storages\ChangeInfoFactory::createTermChangeInfo'
 
 termmeta:
   since: 35329
   id: meta_id
+  changeinfo-fn: 'VersionPress\Storages\ChangeInfoFactory::createTermMetaChangeInfo'
   parent-reference: term_id
   references:
     term_id: term
 
 term_taxonomy:
   id: term_taxonomy_id
+  changeinfo-fn: 'VersionPress\Storages\ChangeInfoFactory::createTermTaxonomyChangeInfo'
   references:
     parent: term
     term_id: term
@@ -81,6 +91,7 @@ term_taxonomy:
 usermeta:
   id: umeta_id
   parent-reference: user_id
+  changeinfo-fn: 'VersionPress\Storages\ChangeInfoFactory::createUserMetaChangeInfo'
   references:
     user_id: user
   ignored-entities:
@@ -92,6 +103,7 @@ usermeta:
 option:
   table: options
   vpid: option_name
+  storage: 'VersionPress\Storages\OptionStorage'
   value-references:
     option_name@option_value:
       site_icon: post
