--- plugins/versionpress/src/Storages/UserStorage.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace VersionPress\Storages;
-
-use VersionPress\ChangeInfos\UserChangeInfo;
-
-class UserStorage extends DirectoryStorage
-{
-
-    protected function createChangeInfo($oldEntity, $newEntity, $action = null)
-    {
-        return new UserChangeInfo($action, $newEntity["vp_id"], $newEntity["user_login"]);
-    }
-}
